refactor(media): relocate music and voice modules
