+++ tests/media/test_media_module_compatibility.py
@@ -1,0 +1,22 @@
+from __future__ import annotations
+
+from importlib import import_module
+
+import pytest
+
+
+@pytest.mark.parametrize(
+    ("legacy_name", "canonical_name"),
+    (
+        ("music_reply", "runtime.media.music_reply"),
+        ("voice_direction", "runtime.media.voice_direction"),
+    ),
+)
+def test_legacy_media_module_is_the_canonical_module(
+    legacy_name: str,
+    canonical_name: str,
+) -> None:
+    legacy_module = import_module(legacy_name)
+    canonical_module = import_module(canonical_name)
+
+    assert legacy_module is canonical_module
